Renames exception module to exceptions
