+++ neomodel/exceptions.py
@@ -1,6 +1,7 @@
 import importlib
 
 
+
 class AttemptedCardinalityViolation(Exception):
     """
     Attempted to alter the database state against the cardinality definitions.
